🔨 getProConDiscussions & Detail 생성 - #101
